# server/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Dispute2Suit REST API
  description: |
    The Dispute2Suit REST API v1 provides programmatic access to client, case, document,
    and dashboard data for authorized partners. Read endpoints retrieve data, while write endpoints
    allow partners to add comments and update case statuses. Every request is scoped to the partner
    associated with the API key, ensuring strict data isolation between organizations.

    ## Authentication

    All endpoints require a valid API key passed in the `Authorization` header using the Bearer scheme.
    API keys are generated by an administrator through the **API Keys** page in the admin dashboard
    and are tied to a specific partner organization.

    ```
    Authorization: Bearer d2s_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    ## Rate Limiting

    Each API key is allowed **100 requests per 15-minute window**. When the limit is exceeded,
    the API returns `429 Too Many Requests` until the window resets.

    ## Data Scoping

    All data is automatically scoped to the partner organization associated with the API key.
    A partner cannot access resources belonging to another partner. Requests for out-of-scope
    resources return `404 Not Found` to prevent information leakage.

    ## Data Sanitization

    Sensitive fields (SSN, internal storage keys, IP addresses, user agent strings, internal user IDs)
    are automatically stripped from all responses.
  version: 1.0.0
  contact:
...
  license:
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        API key in the format `d2s_` followed by 40 random characters.
        Pass as a Bearer token in the Authorization header.
  schemas:
...
    DashboardStats:
      type: object
      properties:
        totalClients:
          type: integer
          description: Total number of clients for this partner
          example: 12
        clientsByStage:
          type: object
          description: Client count grouped by stage
          additionalProperties:
            type: integer
          example:
            lead: 3
            prospect: 2
            active: 5
            completed: 1
            inactive: 1
        totalCases:
          type: integer
          description: Total number of cases
          example: 28
        activeCases:
          type: integer
          description: Number of cases in active statuses
          example: 18
        completedCases:
          type: integer
          description: Number of cases in completed statuses (settled, settlement_paid_out, closed)
          example: 7
        rejectedCases:
          type: integer
          description: Number of rejected cases
          example: 3
        casesByStatus:
          type: object
          description: Case count grouped by status
          additionalProperties:
            type: integer
          example:
            new: 5
            pending_review: 4
            in_review: 3
            settled: 4
        casesByPriority:
          type: object
          description: Case count grouped by priority
          additionalProperties:
            type: integer
          example:
            low: 5
            medium: 12
            high: 8
            urgent: 3
        successRate:
          type: integer
          description: Percentage of completed cases (rounded to nearest integer)
          example: 25
...
  responses:
    Unauthorized:
      description: Authentication failed — missing, malformed, or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_header:
              summary: No Authorization header
              value:
                success: false
                error: "Missing or invalid Authorization header. Use: Bearer <api_key>"
            invalid_format:
              summary: Key does not start with d2s_
              value:
                success: false
                error: Invalid API key format.
            invalid_key:
              summary: Key not found or hash mismatch
              value:
                success: false
                error: Invalid API key.
    Forbidden:
      description: Key is revoked, expired, or lacks the required permission scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            revoked:
              summary: API key has been revoked
              value:
                success: false
                error: API key has been revoked.
            expired:
              summary: API key has expired
              value:
                success: false
                error: API key has expired.
            insufficient_permissions:
              summary: Missing required permission scope
              value:
                success: false
                error: "Insufficient permissions. Required: cases:read"
...
    RateLimited:
      description: Rate limit exceeded (100 requests per 15-minute window)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Rate limit exceeded. Please try again later.
    InternalError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal server error.
security:
  - ApiKeyAuth: []
paths:
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      summary: Get dashboard statistics
      description: |
        Returns aggregate statistics for the authenticated partner's clients and cases.
        Provides a high-level overview suitable for dashboard widgets and summary reports.

        **Permission required:** `dashboard:read`
      tags:
        - Dashboard
      responses:
        '200':
          description: Dashboard statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /clients:
    get:
      operationId: listClients
      summary: List clients
      description: |
        Returns a paginated list of clients belonging to the authenticated partner.
        Supports filtering by stage and searching by name or email.
        Client records are sanitized to exclude SSN and custom fields.

        **Permission required:** `clients:read`
      tags:
        - Clients
      parameters:
        - name: page
          in: query
          description: Page number (starts at 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Records per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: stage
          in: query
          description: Filter by client stage
          schema:
            type: string
            enum: [lead, prospect, active, completed, inactive]
        - name: search
          in: query
          description: Case-insensitive search across client name and email
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of clients
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Client'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /clients/{id}:
    get:
      operationId: getClient
      summary: Get client details
      description: |
        Returns detailed information for a single client, including an array of their
        associated cases (scoped to the authenticated partner). The client must belong
        to the authenticated partner; otherwise, a `404` is returned.

        **Permission required:** `clients:read`
      tags:
        - Clients
      parameters:
        - name: id
          in: path
          required: true
          description: Client numeric ID
          schema:
            type: integer
      responses:
        '200':
          description: Client details with associated cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ClientDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /cases:
    get:
      operationId: listCases
      summary: List cases
      description: |
        Returns a paginated list of cases belonging to the authenticated partner.
        Supports filtering by status, priority, client ID, and searching by case title.

        **Permission required:** `cases:read`
      tags:
        - Cases
      parameters:
        - name: page
          in: query
          description: Page number (starts at 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Records per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: Filter by case status
          schema:
            type: string
            enum:
              - new
              - pending_review
              - in_review
              - more_info_needed
              - ready_for_attorney
              - sent_to_attorney
              - accepted_by_attorney
              - rejected
              - settled
              - settlement_paid_out
              - closed
        - name: priority
          in: query
          description: Filter by case priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: clientId
          in: query
          description: Filter cases for a specific client
          schema:
            type: integer
        - name: search
          in: query
          description: Case-insensitive search across case titles
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /cases/{id}:
    get:
      operationId: getCase
      summary: Get case details
      description: |
        Returns detailed information for a single case, including its associated tasks
        and documents. Internal fields (`createdBy` on tasks, `uploadedBy` and `fileKey`
        on documents) are stripped from the response.

        **Permission required:** `cases:read`
      tags:
        - Cases
      parameters:
        - name: id
          in: path
          required: true
          description: Case numeric ID
          schema:
            type: integer
      responses:
        '200':
          description: Case details with tasks and documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/CaseDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /cases/{id}/timeline:
    get:
      operationId: getCaseTimeline
      summary: Get case timeline
      description: |
        Returns the activity timeline for a specific case. Timeline entries represent
        status changes, comments, document uploads, and other auditable events.
        Sensitive audit fields (`ipAddress`, `userAgent`, `userId`) are stripped.

        **Permission required:** `cases:read`
      tags:
        - Cases
      parameters:
        - name: id
          in: path
          required: true
          description: Case numeric ID
          schema:
            type: integer
      responses:
        '200':
          description: Case timeline entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimelineEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /cases/{id}/documents:
    get:
      operationId: getCaseDocuments
      summary: Get case documents
      description: |
        Returns all documents associated with a specific case. Internal storage fields
        (`fileKey`, `uploadedBy`) are excluded from the response.

        **Permission required:** `documents:read`
      tags:
        - Cases
      parameters:
        - name: id
          in: path
          required: true
          description: Case numeric ID
          schema:
            type: integer
      responses:
        '200':
          description: List of case documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /cases/{id}/comments:
    get:
      operationId: getCaseComments
      summary: Get case comments
      description: |
        Returns all comments on a specific case. The internal `userId` field
        is stripped from each comment record.

        **Permission required:** `cases:read`
      tags:
        - Cases
      parameters:
        - name: id
          in: path
          required: true
          description: Case numeric ID
          schema:
            type: integer
      responses:
        '200':
          description: List of case comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: addCaseComment
      summary: Add a comment to a case
      description: |
        Creates a new comment on a specific case. The comment is attributed to the
        partner organization associated with the API key, with the author name
        displayed as `"<Partner Name> (API)"`.

        An activity log entry is automatically created to record the action.

        **Permission required:** `cases:write`
      tags:
        - Cases
      parameters:
        - name: id
          in: path
          required: true
          description: Case numeric ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
            example:
              comment: Client provided additional documentation for the dispute.
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Comment'
              example:
                success: true
                data:
                  id: 5
                  caseId: 3
                  userName: "Test Law Firm LLC (API)"
                  userRole: partner
                  comment: Client provided additional documentation for the dispute.
                  createdAt: "2026-03-28T16:00:00.000Z"
        '400':
          description: Validation error (missing or invalid comment)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_comment:
                  summary: Comment field is missing or empty
                  value:
                    success: false
                    error: "Field 'comment' is required and must be a non-empty string."
                too_long:
                  summary: Comment exceeds 5000 characters
                  value:
                    success: false
                    error: Comment must not exceed 5000 characters.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...